Created reserve and revert stock endpoints with pessimistic locking logic, updated ingridient and meal entities to address the many to many relationship problem

+++ menu-management-service/src/main/java/com/restaurant/menu/dto/BulkUpdateStockRequest.java
@@ -17,6 +17,6 @@ public record BulkUpdateStockRequest(
 
             @NotNull(message = "Stock quantity is required")
             @PositiveOrZero(message = "Stock cannot be negative")
-            Integer stock) {
+            Double stock) { // in grams
     }
 }
