+++ menu-management-service/src/main/java/com/restaurant/menu/dto/UpdateStockRequest.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.PositiveOrZero;
 public record UpdateStockRequest(
         @NotNull(message = "Stock quantity is required")
         @PositiveOrZero(message = "Stock cannot be negative")
-        Integer stock) {
+        Double stock) { // in grams
 }
