+++ menu-management-service/src/main/java/com/restaurant/menu/entity/Ingredient.java
@@ -34,7 +34,7 @@ public class Ingredient {
 
     @Builder.Default
     @Column(name = "stock", nullable = false)
-    private Integer stock = 0;
+    private Double stock = 0.0; // stored in grams
 
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
